+++ Assets/Temporary/hiramaTemporary/Scripte/Enemy.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: fcd8292f9de3b034b94d3909e59c52a7
+guid: 90ee3cf7877aaad4e99f2946dac1e07b
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
