+++ Assets/Temporary/hiramaTemporary/Scripte/Move.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 90ee3cf7877aaad4e99f2946dac1e07b
+guid: 965c30f099b23f14bbe1b3e74be73e6b
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
